+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.3
-* added `is_slash_term()` global function.
+* added `clean_url()` and `is_slash_term()` global functions.
 
 ### 1.0.2
 * added `Tools\Apache` class with some methods;